// frontend/libs/canvasSpreadsheet/src/lib/types/types.ts
export type GridApi = {
  getViewportCoords: () => ViewportCoords;
  getViewportEdges: () => ViewportEdges;
  moveViewport: (x: number, y: number) => void;
  moveViewportToCell: (
    col: number,
    row: number,
    centerCellInViewport?: boolean,
  ) => void;

  updateSelection: (
    selection: SelectionEdges | null,
    selectionOptions?: SelectionOptions,
  ) => void;
  clearSelection: () => void;
  selectedTable: string | null;
  selectedChart: string | null;
  setSelectedChart: (chartName: string | null) => void;

  getCell: GetCell;
  getCellFromCoords: (x: number, y: number) => CellPlacement;
  getCellX: GetCellX;
  getCellY: GetCellY;
  gridSizes: GridSizes;
  getCanvasSymbolWidth: () => number;
  getColumnContentMaxSymbols: (
    col: number,
    viewportStartRow: number,
    viewportEndRow: number,
  ) => number;

  setCellValue: (col: number, row: number, value: string) => void;

  gridViewportSubscription: (callback: () => void) => () => void;

  tooltipEvent$: Subject<GridTooltipEvent>;
  openTooltip: (x: number, y: number, message: string) => void;
  closeTooltip: () => void;

  contextMenuEvent$: Subject<GridContextMenuEvent>;
  openContextMenuAtCoords: (
    x: number,
    y: number,
    col: number,
    row: number,
    source?: 'canvas-element' | 'html-element',
  ) => void;

  cellEditorEvent$: Subject<GridCellEditorEvent>;
  isCellEditorOpen: () => boolean;
  isCellEditorFocused: () => boolean;
  hideCellEditor: () => void;
  setCellEditorValue: (value: string) => void;
  insertCellEditorValue: (
    value: string,
    options?: GridCellEditorEventInsertValue['options'],
  ) => void;
  showCellEditor: (
    col: number,
    row: number,
    value: string,
    options?: CellEditorExplicitOpenOptions,
  ) => void;

  setPointClickValue: (value: string) => void;
  setPointClickError: (error: boolean) => void;

  hideDottedSelection: () => void;
  showDottedSelection: (selection: SelectionEdges) => void;

  dndSelection: SelectionEdges | null;
  setDNDSelection: (selection: SelectionEdges | null) => void;

  arrowNavigation: (key: string) => void;
  tabNavigation: () => void;

  isPanModeEnabled: boolean;

  hasCharts: boolean;
  setHasCharts: (hasCharts: boolean) => void;

  events$: Observable<EventType>;

  event: any;
  updateSelectionAfterDataChanged: (selection: SelectionEdges) => void;

  selection$: BehaviorSubject<Edges | null>;
};

export type GetCellX = (col: number) => number;
export type GetCellY = (row: number) => number;
export type GetCell = (col: number, row: number) => GridCell | undefined;

export type CanvasOptions = {
  enableMoveTable?: boolean;
  enableOverflowComponents?: boolean;
  placement: 'body' | 'modal';
  showDNDSelection?: boolean;
  showDottedSelection?: boolean;
  showErrors?: boolean;
  showExpandButton?: boolean;
  showHiddenCells?: boolean;
  showNotes?: boolean;
  showOverrides?: boolean;
  showResizers?: boolean;
  showSelection?: boolean;
  showTableBorders?: boolean;
  showWelcomeMessage?: boolean;
  colNumberType: 'number' | 'excel';
};

export type SheetControl = {
  tableName: string;
  fieldName: string;
  controlType: ControlType;
  controlSourcesTables: string[];
};

export type GridProps = {
  canvasId: string;
  data: GridData;
  chartData?: ChartsData;
  filterList?: GridListFilter[];
  sheetControls?: SheetControl[];
  tableStructure: GridTable[];
  columnSizes?: Record<string, number>;
  inputFiles?: (ResourceMetadata | SharedWithMeMetadata)[] | null;
  controlData?: ControlData | null;
  controlIsLoading?: boolean;

  isReadOnly?: boolean;
  zoom?: number;
  functions?: FunctionInfo[];
  parsedSheets?: ParsedSheets;
  charts?: GridChart[];
  theme: AppTheme;
  formulaBarMode?: FormulaBarMode;
  isPointClickMode?: boolean;
  sheetContent?: string;
  systemMessageContent?: GPTState;
  currentSheetName?: string | null;
  viewportInteractionMode?: ViewportInteractionMode;
  showGridLines?: boolean;

  eventBus: GridEventBus;

  canvasOptions?: CanvasOptions;
};
// ...
export type Edges = {
  startRow: number;
  endRow: number;
  startCol: number;
  endCol: number;
};

export type ViewportEdges = Edges;
export type SelectionEdges = Edges;

export type ViewportCoords = {
  x1: number;
  y1: number;
  x2: number;
  y2: number;
};
// ...
export type SelectionOptions = {
  selectedTable?: string;
  silent?: boolean;
};

export type GridTable = {
  startRow: number;
  startCol: number;
  endCol: number;
  endRow: number;
  tableName: string;
  chartType?: ChartType;
  isTableNameHeaderHidden: boolean;
  isTableFieldsHeaderHidden: boolean;
  isTableHorizontal: boolean;
  totalSize: number;
  hasKeys: boolean;
  isManual: boolean;
  highlightType?: 'DIMMED' | 'NORMAL' | 'HIGHLIGHTED' | undefined;
  note?: string;
  fieldNames: string[];
};
// ...
export type GridCell = {
  table?: GridTable;
  field?: GridField;
  value?: string;
  displayValue?: string;

  hasError?: boolean;
  errorMessage?: string;

  totalIndex?: number;
  totalType?: TotalType;
  totalExpression?: string;

  row: number;
  col: number;
  dataIndex?: number;

  overrideIndex?: number;
  overrideValue?: OverrideValue;
  isOverride?: boolean;
  overrideAIFunctions?: boolean;

  isUrl?: boolean;

  isTableHeader?: boolean;
  isFieldHeader?: boolean;

  isRightAligned?: boolean;

  startCol: number;
  endCol: number;

  startGroupColOrRow: number;
  endGroupColOrRow: number;
};

export type RowData = { [col: string]: GridCell };

export type GridData = {
  [row: string]: RowData;
};